Estructura base configuración de subtipos

+++ database/migrations/2022_10_10_200639_create_table_contract_configurations.php
@@ -17,7 +17,7 @@ class CreateTableContractConfigurations extends Migration
 
             $table->bigIncrements('id');
 
-            $table->unsignedInteger('fk_id_configuration_subtype');
+            $table->unsignedBigInteger('fk_id_configuration_subtype');
             $table->foreign('fk_id_configuration_subtype','fk_contract_to_subtype')->references('id')->on('fac_configuration_subtypes');
 
             $table->unsignedInteger('fk_id_activity')->nullable();
